add email and email files policies

+++ app/Http/Controllers/Admin/Notify/EmailController.php
@@ -15,6 +15,7 @@ class EmailController extends Controller
      */
     public function index()
     {
+        $this->authorize('index', Email::class);
         $emails = Email::orderBy('created_at', 'desc')->simplePaginate(15);
         return view('admin.notify.email.index', compact('emails'));
     }
@@ -26,6 +27,7 @@ class EmailController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', Email::class);
         return view('admin.notify.email.create');
     }
 
@@ -37,7 +39,8 @@ class EmailController extends Controller
      */
     public function store(EmailRequest $request)
     {
-        
+        $this->authorize('create', Email::class);
+
         $inputs = $request->all();
         $inputs['published_at'] = date('Y-m-d H:i:s', intval(substr($request->published_at, 0, 10)));
         $inputs['author_id'] = 1;
@@ -64,6 +67,7 @@ class EmailController extends Controller
      */
     public function edit(Email $email)
     {
+        $this->authorize('update', Email::class);
         return view('admin.notify.email.edit', compact('email'));
     }
 
@@ -76,6 +80,7 @@ class EmailController extends Controller
      */
     public function update(EmailRequest $request, Email $email)
     {
+        $this->authorize('update', Email::class);
         $inputs = $request->all();
         $inputs['published_at'] = date('Y-m-d H:i:s', intval(substr($request->published_at, 0, 10)));
         $email->update($inputs);
@@ -90,23 +95,24 @@ class EmailController extends Controller
      */
     public function destroy(Email $email)
     {
+        $this->authorize('delete', Email::class);
         $email->delete();
         return redirect()->route('admin.notify.email.index')->with('alertify-error', 'ایمیل موردنظر با موفقیت حذف شد.');
     }
 
     public function status(Email $email)
     {
+        $this->authorize('update', Email::class);
+
         $email->status = $email->status == 0 ? 1 : 0;
         $result = $email->save();
 
         if ($result) {
             if ($email->status == 0)
-            return response()->json(['status' => true, 'checked' => false]);
+                return response()->json(['status' => true, 'checked' => false]);
             else
-            return response()->json(['status' => true, 'checked' => true]);
-        }
-        else
-        {
+                return response()->json(['status' => true, 'checked' => true]);
+        } else {
             return response()->json(['status' => false]);
         }
     }
